contenu

+++ database/migrations/2023_12_07_000000_add_columns_to_user_addresses_table.php
@@ -24,7 +24,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('user_addresses', function (Blueprint $table) {
-            $table->dropColumn('vendor_id');
+            if (Schema::hasColumn('user_addresses', 'vendor_id')) {
+                $table->dropColumn('vendor_id');            
+			}
       });   
     }
 };
